Перенес служебные ф-ции в utils.js
