+++ index.html
@@ -230,7 +230,7 @@
     </div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script src="./js/main.js" type="module"></script>
 
 </body>
 </html>
